7주차 - SearchBar 컴포넌트 분리
