+++ src/pages/ContentsArea.js
@@ -1,19 +1,6 @@
 import Card from '../components/Card';
-import searchIcon from '../assets/icon-search.svg';
+import SearchBar from '../components/SearchBar';
 import './ContentsArea.css';
-
-const SearchBar = () => {
-  return (
-    <form className="search_link">
-      <img src={searchIcon} alt="search" />
-      <input
-        type="text"
-        className="input_box"
-        placeholder="링크를 검색해 보세요."
-      />
-    </form>
-  );
-};
 
 const ContentsArea = ({ links }) => {
   return (
